requirements.txt
flask
flask-Restful
pyjwt
flask-sqlalchemy
werkzeug